feat: add dependency liquibase in pom.xml, update application.properties, db.changelog-master.xml

+++ src/main/java/ru/tracker/calorie_tracker/repository/DishRepository.java
@@ -3,5 +3,5 @@ package ru.tracker.calorie_tracker.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.tracker.calorie_tracker.model.Dish;
 
-public interface DishRepository extends JpaRepository<Dish, Integer> {
+public interface DishRepository extends JpaRepository<Dish, Long> {
 }
